cmsis: Add CMSIS-Core(R)
This commit adds a preliminary implementation of CMSIS-Core(R) for use
with the ARM Cortex-R port.

At this time, CMSIS-Core(R) is not merged back into the upstream CMSIS
repository and therefore is not available from official sources.

Until upstream merge happens, the preliminary version can be obtained
from the following URL:

https://github.com/stephanosio/CMSIS_5/tree/core_r

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>
Signed-off-by: Carles Cufi <carles.cufi@nordicsemi.no>