cmsis: add top-level header file for Cortex-A9 CPUs
Add a top-level header file for Cortex-A9 CPUs, which enables MMU
and FPU support in the Cortex-A common peripheral access layer.
It is based on the equivalent header file for the Cortex-R5.

This addition is required for proper use of CMSIS abstraction in
the aarch32 MMU code. Comp. Zephyr PR 36983.

Signed-off-by: Immo Birnbaum <Immo.Birnbaum@Weidmueller.com>