DSP: Do not generate warnings when scalar functions are used with MVE
This commit updates the DSP functions to print out a compiler message
(non-warning) instead of a warning when the scalar version of a
function implementation is used in spite of the M-Profile Vector
Extension being enabled.

The rationale is that warnings may be treated as errors under certain
circumstances (e.g. when running tests), and the scalar function
implementations being used should not cause a buiuld failure.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>